Small bug fix in seasonal aggregation, plus re-running all
Bug was that the last day in the aggregation period wasn't included in the aggregation. Fixed, and the same function is now used to seasonally aggregate throughout. Fixing it meant everything had to be re-run, discretization boundaries (from regression trees) changed a little bit, as did the fitted BN params, etc. All re-done now, with a little bit of tidying of the notebooks along the way.